#636 - update back-end tests for all change request creation endpoints
